remove undifned for missing field
